microservices/chat_service/requirements.txt
ipython==8.12.0
openai
pinecone-client
python-dotenv==1.0.0
uuid
pathlib
uvicorn
fastapi
jinja2
slowapi
httpx
git+https://github.com/casics/nostril.git
tiktoken
bs4
apscheduler